Added tests for edit_question endpoint
This commit adds unit tests for the edit_question endpoint in admin.py
In order to create tests in line with other tests in the suite, edit_question needed to be modified to add `json.dumps()` calls surrounding all outputs. As otherwise `json.loads()` would fail.
This change only applies to edit_question and should not affect the usage of the interface.