Missing js files for PreTeXt books
The admin interface gets these files from each book’s _static folder.  But they are not used by any of the components or they would be in webpack.  So just put them in static here.

I’ve already weeded out a few obvious files that are obsolete and there are probably more.